+++ internal/default/config/const.go
