+++ internal/default/infrastructure/entity/base.go
