+++ internal/default/usecase/user.go
@@ -3,7 +3,7 @@ package usecase
 import (
 	"context"
 
-	"github.com/abyssparanoia/rapid-go/default/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/default/domain/model"
 )
 
 // User ... inteface of User usecase
