+++ internal/default/usecase/user_impl.go
@@ -3,8 +3,8 @@ package usecase
 import (
 	"context"
 
-	"github.com/abyssparanoia/rapid-go/default/domain/model"
-	"github.com/abyssparanoia/rapid-go/default/domain/repository"
+	"github.com/abyssparanoia/rapid-go/internal/default/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/default/domain/repository"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/log"
 )
